refactor: Updating webhook routes

+++ src/Admin/WebhookRegister.php
@@ -57,7 +57,11 @@ class WebhookRegister {
 	 */
 	public function register() {
 		Util::validate_user_permissions();
-		Util::validate_nonce( 'mondu_trade_register_webhooks', 'mondu_trade_register_webhooks_nonce' );
+
+		$is_nonce_valid = wp_verify_nonce( 'mondu_trade_register_webhooks_nonce', 'mondu_trade_register_webhooks' );
+		if ( ! $is_nonce_valid ) {
+			Util::die_after_security_check();
+		}
 
 		try {
 			$webhooks = $this->mondu_request_wrapper->get_webhooks();
@@ -112,8 +116,14 @@ class WebhookRegister {
 			exit;
 		}
 
-		$uuid = isset( $_POST['uuid'] ) ? sanitize_text_field( wp_unslash( $_POST['uuid']) ) : '';
-		Util::validate_nonce( 'mondu_trade_delete_webhook_' . $uuid, 'mondu_trade_delete_webhook_nonce' );
+		$uuid = isset( $_POST['uuid'] ) ? sanitize_text_field( wp_unslash( $_POST['uuid'] ) ) : '';
+
+		$is_nonce_valid = wp_verify_nonce( 'mondu_trade_delete_webhook_nonce', 'mondu_trade_delete_webhook_' . $uuid, );
+		if ( ! $is_nonce_valid ) {
+			Util::die_after_security_check();
+
+			return;
+		}
 
 		try {
 			$this->mondu_request_wrapper->api->delete_webhook( [ 'uuid' => $uuid ] );
@@ -145,10 +155,12 @@ class WebhookRegister {
 			'webhook_delete_failed'   => __( 'Failed to delete webhook.', 'mondu-trade-account' ),
 		];
 
+		// phpcs:disable WordPress.Security.NonceVerification.Recommended
 		$message_key = self::ADMIN_MESSAGE_KEY;
-		$message = isset( $_GET[ $message_key ] )
+		$message     = isset( $_GET[ $message_key ] )
 			? sanitize_text_field( wp_unslash( $_GET[ $message_key ] ) )
 			: '';
+		// phpcs:enable
 
 		if ( $message && isset( $messages[ $message ] ) ) {
 			return $messages[ $message ];
